+++ src/words/ast/INodeParameter.java
@@ -7,10 +7,19 @@ public class INodeParameter extends INode {
 	public INodeParameter(Object... children) {
 		super(children);
 	}
-
+	
 	@Override
 	public ASTValue eval(Environment environment) throws WordsRuntimeException {
-		// TODO
-		throw new AssertionError("Not yet implemented");
+		assert false: "Requires an inherited object";
+		return null;
+	}
+
+	@Override
+	public ASTValue eval(Environment environment, Object inherited) throws WordsRuntimeException {
+		CustomAction customAction = (CustomAction) inherited;
+		ASTValue param = children.get(0).eval(environment);
+		assert param.type == ASTValue.Type.STRING : "Parameter names should always be strings";
+		customAction.addParameter(param.stringValue);
+		return null;
 	}
 }
