+++ system_test/CustomActionTest.words.log
@@ -1,0 +1,300 @@
+frame #: 1
+(0,0):
+<person> Fred
+frame #: 2
+(0,0):
+<person> Fred "I will jump"
+frame #: 3
+(0,0):
+<person> Fred "I'm jumping!"
+frame #: 4
+(0,1):
+<person> Fred "I'm jumping!"
+frame #: 5
+(0,2):
+<person> Fred "I'm jumping!"
+frame #: 6
+(0,1):
+<person> Fred "I'm jumping!"
+frame #: 7
+(0,0):
+<person> Fred "I'm jumping!"
+frame #: 8
+(0,0):
+<person> Fred "I jumped!"
+frame #: 9
+(0,0):
+<person> Fred "I jumped!"
+frame #: 10
+(0,0):
+<person> Fred "I jumped!"
+frame #: 11
+(0,0):
+<person> Fred "I jumped!"
+frame #: 12
+(0,0):
+<person> Fred "I jumped!"
+frame #: 13
+(0,0):
+<person> Fred "I jumped!"
+frame #: 14
+(0,0):
+<person> Fred "I jumped!"
+frame #: 15
+(0,0):
+<person> Fred "I jumped!"
+frame #: 16
+(0,0):
+<person> Fred "I jumped!"
+frame #: 17
+(0,0):
+<person> Fred "I jumped!"
+frame #: 18
+(0,0):
+<person> Fred "I jumped!"
+frame #: 19
+(0,0):
+<person> Fred "I jumped!"
+frame #: 20
+(0,0):
+<person> Fred "I jumped!"
+frame #: 21
+(0,0):
+<person> Fred "I jumped!"
+frame #: 22
+(0,0):
+<person> Fred "I jumped!"
+frame #: 23
+(0,0):
+<person> Fred "I jumped!"
+frame #: 24
+(0,0):
+<person> Fred "I jumped!"
+frame #: 25
+(0,0):
+<person> Fred "I jumped!"
+frame #: 26
+(0,0):
+<person> Fred "I jumped!"
+frame #: 27
+(0,0):
+<person> Fred "I jumped!"
+frame #: 28
+(0,0):
+<person> Fred "I jumped!"
+frame #: 29
+(0,0):
+<person> Fred "I jumped!"
+frame #: 30
+(0,0):
+<person> Fred "I jumped!"
+frame #: 31
+(0,0):
+<person> Fred "I jumped!"
+frame #: 32
+(0,0):
+<person> Fred "I jumped!"
+frame #: 33
+(0,0):
+<person> Fred "I jumped!"
+frame #: 34
+(0,0):
+<person> Fred "I jumped!"
+frame #: 35
+(0,0):
+<person> Fred "I jumped!"
+frame #: 36
+(0,0):
+<person> Fred "I jumped!"
+frame #: 37
+(0,0):
+<person> Fred "I jumped!"
+frame #: 38
+(0,0):
+<person> Fred "I jumped!"
+frame #: 39
+(0,0):
+<person> Fred "I jumped!"
+frame #: 40
+(0,0):
+<person> Fred "I jumped!"
+frame #: 41
+(0,0):
+<person> Fred "I jumped!"
+frame #: 42
+(0,0):
+<person> Fred "I jumped!"
+frame #: 43
+(0,0):
+<person> Fred "I jumped!"
+frame #: 44
+(0,0):
+<person> Fred "I jumped!"
+frame #: 45
+(0,0):
+<person> Fred "I jumped!"
+frame #: 46
+(0,0):
+<person> Fred "I jumped!"
+frame #: 47
+(0,0):
+<person> Fred "I jumped!"
+frame #: 48
+(0,0):
+<person> Fred "I jumped!"
+frame #: 49
+(0,0):
+<person> Fred "I jumped!"
+frame #: 50
+(0,0):
+<person> Fred "I jumped!"
+frame #: 51
+(0,0):
+<person> Fred "I jumped!"
+frame #: 52
+(0,0):
+<person> Fred "I jumped!"
+frame #: 53
+(0,0):
+<person> Fred "I jumped!"
+frame #: 54
+(0,0):
+<person> Fred "I jumped!"
+frame #: 55
+(0,0):
+<person> Fred "I jumped!"
+frame #: 56
+(0,0):
+<person> Fred "I jumped!"
+frame #: 57
+(0,0):
+<person> Fred "I jumped!"
+frame #: 58
+(0,0):
+<person> Fred "I jumped!"
+frame #: 59
+(0,0):
+<person> Fred "I jumped!"
+frame #: 60
+(0,0):
+<person> Fred "I jumped!"
+frame #: 61
+(0,0):
+<person> Fred "I jumped!"
+frame #: 62
+(0,0):
+<person> Fred "I jumped!"
+frame #: 63
+(0,0):
+<person> Fred "I jumped!"
+frame #: 64
+(0,0):
+<person> Fred "I jumped!"
+frame #: 65
+(0,0):
+<person> Fred "I jumped!"
+frame #: 66
+(0,0):
+<person> Fred "I jumped!"
+frame #: 67
+(0,0):
+<person> Fred "I jumped!"
+frame #: 68
+(0,0):
+<person> Fred "I jumped!"
+frame #: 69
+(0,0):
+<person> Fred "I jumped!"
+frame #: 70
+(0,0):
+<person> Fred "I jumped!"
+frame #: 71
+(0,0):
+<person> Fred "I jumped!"
+frame #: 72
+(0,0):
+<person> Fred "I jumped!"
+frame #: 73
+(0,0):
+<person> Fred "I jumped!"
+frame #: 74
+(0,0):
+<person> Fred "I jumped!"
+frame #: 75
+(0,0):
+<person> Fred "I jumped!"
+frame #: 76
+(0,0):
+<person> Fred "I jumped!"
+frame #: 77
+(0,0):
+<person> Fred "I jumped!"
+frame #: 78
+(0,0):
+<person> Fred "I jumped!"
+frame #: 79
+(0,0):
+<person> Fred "I jumped!"
+frame #: 80
+(0,0):
+<person> Fred "I jumped!"
+frame #: 81
+(0,0):
+<person> Fred "I jumped!"
+frame #: 82
+(0,0):
+<person> Fred "I jumped!"
+frame #: 83
+(0,0):
+<person> Fred "I jumped!"
+frame #: 84
+(0,0):
+<person> Fred "I jumped!"
+frame #: 85
+(0,0):
+<person> Fred "I jumped!"
+frame #: 86
+(0,0):
+<person> Fred "I jumped!"
+frame #: 87
+(0,0):
+<person> Fred "I jumped!"
+frame #: 88
+(0,0):
+<person> Fred "I jumped!"
+frame #: 89
+(0,0):
+<person> Fred "I jumped!"
+frame #: 90
+(0,0):
+<person> Fred "I jumped!"
+frame #: 91
+(0,0):
+<person> Fred "I jumped!"
+frame #: 92
+(0,0):
+<person> Fred "I jumped!"
+frame #: 93
+(0,0):
+<person> Fred "I jumped!"
+frame #: 94
+(0,0):
+<person> Fred "I jumped!"
+frame #: 95
+(0,0):
+<person> Fred "I jumped!"
+frame #: 96
+(0,0):
+<person> Fred "I jumped!"
+frame #: 97
+(0,0):
+<person> Fred "I jumped!"
+frame #: 98
+(0,0):
+<person> Fred "I jumped!"
+frame #: 99
+(0,0):
+<person> Fred "I jumped!"
+frame #: 100
+(0,0):
+<person> Fred "I jumped!"
